Base code put in place for level transition
Option 1: Change scenes for each level.

Option 2: Activate new level and deactivate old level, teleport player to entrance.
Change scenes on new floor.

Option 3: Have a door that opens when they stand on a button or get close to the door?
this could also activate the enemies in the next room or similar.
Change scenes on new floor.